fix: moved FetchContent at root, changed includeDir, updated doc
